Move CTeamStrategy traits
